+++ benchmarks/Logging.Performance/ScopesOverheadBenchmark.cs
@@ -1,52 +1,11 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System;
 using BenchmarkDotNet.Attributes;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Microsoft.Extensions.Logging.Performance
 {
-    [AspNetCoreBenchmark]
-    public class LogValuesBenchmarks : LoggingBenchmarkBase
-    {
-        private ILogger _logger;
-
-        [Benchmark]
-        public void NotFiltered_InsideScope()
-        {
-            TwoArgumentErrorMessage(_logger, 1, "string", Exception);
-        }
-
-        [GlobalSetup]
-        public void Setup()
-        {
-            var services = new ServiceCollection();
-            services.AddLogging();
-            services.AddSingleton<ILoggerProvider, LoggerProvider<MessageFormattingLogger>>();
-
-            _logger = services.BuildServiceProvider().GetService<ILoggerFactory>().CreateLogger("Logger");
-        }
-
-        public class MessageFormattingLogger: ILogger
-        {
-            public void Log<TState>(LogLevel logLevel, EventId eventId, TState state, Exception exception, Func<TState, Exception, string> formatter)
-            {
-                formatter(state, exception);
-            }
-
-            public bool IsEnabled(LogLevel logLevel)
-            {
-                return true;
-            }
-
-            public IDisposable BeginScope<TState>(TState state)
-            {
-                return null;
-            }
-        }
-    }
-
     [AspNetCoreBenchmark]
     public class ScopesOverheadBenchmark: LoggingBenchmarkBase
     {
